updated version from Vogsphere

+++ libft/ft_abs.c
@@ -5,14 +5,14 @@
 /*                                                     +:+                    */
 /*   By: rsteigen <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
-/*   Created: 2019/04/10 14:13:58 by rsteigen      #+#    #+#                 */
-/*   Updated: 2019/04/10 14:18:59 by rsteigen      ########   odam.nl         */
+/*   Created: 2019/04/10 14:13:58 by rsteigen       #+#    #+#                */
+/*   Updated: 2020/01/15 10:38:16 by rsteigen      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_abs(int nb)
+long long	ft_abs(long long nb)
 {
 	if (nb < 0)
 		nb = -nb;
